Fixing Configs
