+++ Global/Constant/__init__.py
@@ -1,7 +1,4 @@
 import os
-from Global.Class import Config
 
 ROOT       = os.getcwd()
 CONFIGFILE = '.configfiles/config.ini'
-
-Config_Manager = Config(CONFIGFILE)
